Add interface to receive grpc messages, move node mapper logics out of stub
