+++ internal/core/protocol.go
@@ -1,0 +1,124 @@
+package core
+
+import (
+	"fmt"
+	pb "github.com/media-streaming-mesh/msm-cp/api/v1alpha1/msm_cp"
+	"github.com/media-streaming-mesh/msm-cp/internal/config"
+	"github.com/media-streaming-mesh/msm-cp/internal/model"
+	"github.com/media-streaming-mesh/msm-cp/internal/rtm"
+	"github.com/media-streaming-mesh/msm-cp/internal/stub"
+	node_mapper "github.com/media-streaming-mesh/msm-cp/pkg/node-mapper"
+	stream_mapper "github.com/media-streaming-mesh/msm-cp/pkg/stream-mapper"
+	"github.com/sirupsen/logrus"
+	"google.golang.org/grpc/peer"
+	"io"
+	"net"
+	"sync"
+)
+
+type API interface {
+	Send(conn pb.MsmControlPlane_SendServer) error
+}
+
+type Protocol struct {
+	cfg         *config.Cfg
+	logger      *logrus.Logger
+	stubHandler *stub.StubHandler
+	rtmImpl     rtm.API
+
+	//TODO: move stream_mapper to msm-nc
+	streamMapper *stream_mapper.StreamMapper
+}
+
+func New(cfg *config.Cfg) *Protocol {
+
+	return &Protocol{
+		cfg:          cfg,
+		logger:       cfg.Logger,
+		stubHandler:  stub.NewStubHandler(cfg),
+		streamMapper: stream_mapper.NewStreamMapper(cfg.Logger, new(sync.Map)),
+		rtmImpl:      rtm.New(cfg),
+	}
+}
+
+func (p *Protocol) log(format string, args ...interface{}) {
+	p.logger.Infof("[GRPC] " + fmt.Sprintf(format, args...))
+}
+
+func (p *Protocol) logError(format string, args ...interface{}) {
+	p.logger.Errorf("[GRPC] " + fmt.Sprintf(format, args...))
+}
+
+func (p *Protocol) Send(conn pb.MsmControlPlane_SendServer) error {
+	var ctx = conn.Context()
+	for {
+		// exit if context is done or continue
+		select {
+		case <-ctx.Done():
+			p.log("received connection done")
+			return nil
+		default:
+		}
+
+		//Process stream data
+		stream, err := conn.Recv()
+		if err == io.EOF {
+			// return will close stream-mapper from server side
+			p.logError("found EOF, exiting")
+			return nil
+		}
+		if err != nil {
+			p.logError("received error %v", err)
+			continue
+		}
+
+		var streamData *model.StreamData
+		
+		switch stream.Event {
+		case pb.Event_REGISTER:
+			p.log("Received REGISTER event: %v", stream)
+			//TODO: Find a cleaner way to map node ip for stub
+			contextPeer, _ := peer.FromContext(ctx)
+			remoteAddr, _, _ := net.SplitHostPort(contextPeer.Addr.String())
+			proxyIp, _ := node_mapper.MapNode(remoteAddr)
+			p.stubHandler.OnRegistration(conn, proxyIp)
+
+		case pb.Event_ADD:
+			p.log("Received ADD event: %v", stream)
+			p.rtmImpl.OnAdd(conn, stream)
+			p.stubHandler.OnAdd(conn, stream)
+		case pb.Event_DELETE:
+			p.log("Received DELETE event: %v", stream)
+			streamData, err = p.rtmImpl.OnDelete(stream)
+			p.stubHandler.OnDelete(conn, stream)
+		case pb.Event_DATA:
+			p.log("Received DATA event: %v", stream)
+			streamData, err = p.rtmImpl.OnData(conn, stream)
+		default:
+		}
+
+		if err != nil {
+			return err
+		}
+
+		//Send data to proxy
+		if streamData != nil {
+			//TODO: Writes logical stream graphs to etcd cluster
+			stubAddress := stub.GetStubAddress(streamData.ClientIp, stream.Remote)
+			p.log("StubAddress %v", stubAddress)
+
+			error := p.streamMapper.ProcessStream(model.StreamData{
+				StubIp:      stubAddress,
+				ServerIp:    streamData.ServerIp,
+				ClientIp:    streamData.ClientIp,
+				ServerPorts: streamData.ServerPorts,
+				ClientPorts: streamData.ClientPorts,
+				StreamState: streamData.StreamState,
+			})
+
+			if error != nil {
+				p.logError("ProcessStream failed %v", error)
+			}
+		}
+	}
+}
